method move
