+++ src/utils/index.ts
@@ -1,7 +1,3 @@
-import Web3 from 'web3'
-import { Account } from '@oceanprotocol/lib'
-import { Balance } from '../providers'
-
 export function readFileContent(file: File): Promise<string> {
   return new Promise((resolve, reject) => {
     const reader = new FileReader()
@@ -30,14 +26,4 @@ export const publishFeedback: { [key in number]: string } = {
   4: '4/4 Asset published succesfully'
 }
 
-export async function getAccountId(web3: Web3): Promise<string> {
-  const accounts = await web3.eth.getAccounts()
-  return accounts[0]
-}
-
-export async function getBalance(account: Account): Promise<Balance> {
-  const eth = await account.getEtherBalance()
-  const ocean = await account.getOceanBalance()
-
-  return { eth, ocean }
-}
+export * from './web3'
